# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chismes Api",
        "contact": {
            "email": "soyondori@gmail.com"
        },
        "version": "0.1.0"
    },
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "operationId": "getAuthors",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuthorResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/api/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "operationId": "getAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorResource"
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/api/chismes": {
            "get": {
                "tags": [
                    "Chismes"
                ],
                "operationId": "getChismes",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChismeResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            },
            "post": {
                "tags": [
                    "Chismes"
                ],
                "operationId": "createChisme",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Chisme"
                },
                "responses": {
                    "201": {
                        "description": "Chisme created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChismeResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/api/chismes/{id}": {
            "get": {
                "tags": [
                    "Chismes"
                ],
                "operationId": "getChisme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChismeResource"
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AuthorCommentResource": {
                "title": "Author Comment",
                "description": "Author comment model",
                "required": [
                    "id",
                    "content",
                    "author_id",
                    "recipient_id"
                ],
                "properties": {
                    "id": {
                        "description": "Class AuthorCommentResource",
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "author_id": {
                        "type": "string"
                    },
                    "recipient_id": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Author Comment"
                }
            },
            "AuthorResource": {
                "title": "Author",
                "description": "Author model",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "Class AuthorResource",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "chismes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChismeResource"
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuthorCommentResource"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Author"
                }
            },
            "ChismeCommentResource": {
                "title": "Chisme Comment",
                "description": "Chisme comment model",
                "required": [
                    "id",
                    "content",
                    "author_id",
                    "chisme_id"
                ],
                "properties": {
                    "id": {
                        "description": "Class ChismeCommentResource",
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "author_id": {
                        "type": "string"
                    },
                    "chisme_id": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Chisme Comment"
                }
            },
            "ChismeResource": {
                "title": "Chisme",
                "description": "Chisme model",
                "required": [
                    "id",
                    "title",
                    "content",
                    "author_id"
                ],
                "properties": {
                    "id": {
                        "description": "Class ChismeResource",
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "author_id": {
                        "type": "string"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChismeCommentResource"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Chisme"
                }
            },
            "UserResource": {
                "title": "User",
                "description": "User model",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "Class UserResource",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "requestBodies": {
            "Chisme": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "content": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Chismes",
            "description": "API Endpoints of Chismes"
        },
        {
            "name": "Authors",
            "description": "API Endpoints of Authors"
        }
    ]
}